% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_altname}
\alias{id_altname}
\title{ULAN id alternative names}
\format{An object of class \code{data.frame} with 362779 rows and 2 columns.}
\usage{
id_altname
}
\description{
A data frame with two variables.
}
\keyword{datasets}

